I set up my new testing files and imported/exported accordingly, also set up my new TDD test ready to populate it
